Introduce elasticapmconnector  (#367)
* Move sharedcomponent to /internal

* Introduce elasticapmconnector

This is a new connector that encapsulates:
 - signaltometricsconnector with an Elastic APM-specific configuration
 - lsmintervalprocessor with an Elastic APM-specific configuration

The intention behind this is to simplify EDOT Collector configuration
so users have less configuration to deal with, and less opportunity for
configuration drift and breaking assumptions.

* Apply suggestions from code review

Co-authored-by: Christos Kalkanis <christos.kalkanis@elastic.co>
Co-authored-by: Vishal Raj <vishal.raj@elastic.co>

* Update connector/elasticapmconnector/doc.go

Co-authored-by: Christos Kalkanis <christos.kalkanis@elastic.co>

* Address review comments

* Switch to explicit bucket histograms & fixes

Temporarily switch to explicit bucket histograms for
latency distributions while we work on improving the
performance of exponential histogram merging.

Use the correct histogram definition consistently for
`transaction.duration.summary`.

Fix tests so they properly concatenate metrics before
comparing, and add another resource attribute to the
transaction input so we can check both service_transaction
and transaction metrics.

---------

Co-authored-by: Christos Kalkanis <christos.kalkanis@elastic.co>
Co-authored-by: Vishal Raj <vishal.raj@elastic.co>