Workaround TSDB and timestamp millis precision in mapping
Workaround: If loadgenreceiver generate data points too quickly,
it is possible that 2 data points of the same metric have timestamps of different nanoseconds (precision sent by elasticsearchexporter)
but the same millisecond (precision in ES index mapping).
This is a problem as it will cause TSDB version_conflict_engine_exception.
Set an original timestamp dimension to ensure that these data points have different dimensions
and will not be seen as duplicates by TSDB.