processor(elasticapm): route error to data stream (#938)
We add logic to set the `data_stream.dataset` field to "apm.error" for both
otlp and intake/v2 data going through the processor.

The field can be set in record-level as well as scope/resource level attributes
because the ES exporter searches for it from inner to outer container,
stopping at first match.

Details:

* error events from APM format routed as logs
* span events containing errors routed as traces

---------

Signed-off-by: inge4pres <francesco.gualazzi@elastic.co>